requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.2.0
structlog>=24.1.0
